requirements.txt
flask==3.1.0
flask-cors==4.0.0
flask-socketio==5.5.1
python-dotenv==1.0.0
requests==2.32.3
python-socketio==5.12.0
gunicorn==23.0.0
Werkzeug==3.1.3
psutil==5.9.8
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
langdetect==1.0.9
faker==21.0.0